Experimental support for annotated component types
Right now, component parameters must be specified with default values in
a fields map. It's been suggested that annotated fields on the component
works without changing the interface since the fields map is immutable -
i.e. we don't need to propogate attempts to set values on the map. This
may also allow future versions to perform some level of type checking on
component parameter values.